refactor: delete first version
- Remove first GamblingScreen.
- Rename GamblingScreen2 as GamblingScreen.
