+++ app/src/main/java/com/github/se/cyrcle/ui/gambling/GamblingScreen.kt
@@ -1,302 +1,369 @@
 package com.github.se.cyrcle.ui.gambling
 
-import androidx.compose.animation.AnimatedVisibility
-import androidx.compose.animation.core.*
-import androidx.compose.animation.expandVertically
-import androidx.compose.animation.fadeIn
-import androidx.compose.foundation.Canvas
+import android.content.Context
+import android.graphics.*
+import android.util.AttributeSet
+import android.util.Log
+import android.view.View
 import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.material3.Button
-import androidx.compose.material3.Scaffold
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.scale
-import androidx.compose.ui.geometry.Offset
-import androidx.compose.ui.geometry.Rect
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.Path
-import androidx.compose.ui.graphics.StrokeCap
-import androidx.compose.ui.graphics.drawscope.Stroke
-import androidx.compose.ui.graphics.drawscope.rotate
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.platform.testTag
+import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import kotlinx.coroutines.delay
-import kotlin.math.PI
-import com.github.se.cyrcle.model.user.UserViewModel
-import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.compose.ui.viewinterop.AndroidView
+import kotlin.math.cos
+import kotlin.math.sin
 
-class GamblingScreen {
+class WheelView @JvmOverloads constructor(
+    context: Context,
+    attrs: AttributeSet? = null,
+    defStyleAttr: Int = 0
+) : View(context, attrs, defStyleAttr) {
 
-    private object WheelColors {
-        val Nothing = Color.Gray
-        val Common = Color.Green
-        val Rare = Color(0xFF0066FF)
-        val Epic = Color(0xFF800080)
-        val Legendary = Color(0xFFFFD700)
-    }
+    private val TAG = "WheelView"
 
-    private enum class WheelSegment(val displayName: String) {
-        NOTHING("Nothing"),
-        COMMON("Common Prize!"),
-        RARE("Rare Prize!!"),
-        EPIC("Epic Prize!!!"),
-        LEGENDARY("LEGENDARY!!!!")
-    }
+    private val paint = Paint(Paint.ANTI_ALIAS_FLAG)
+    private val segments = listOf(
+        Segment("Nothing", Color.GRAY),
+        Segment("Legendary", Color.rgb(255, 215, 0)),
+        Segment("Common", Color.GREEN),
+        Segment("Rare", Color.BLUE),
+        Segment("Epic", Color.rgb(128, 0, 128))
+    )
 
-    private fun generateNewPattern(): List<WheelSegment> {
-        return buildList {
-            repeat(65) { add(WheelSegment.NOTHING) }
-            repeat(20) { add(WheelSegment.COMMON) }
-            repeat(10) { add(WheelSegment.RARE) }
-            repeat(4) { add(WheelSegment.EPIC) }
-            repeat(1) { add(WheelSegment.LEGENDARY) }
-        }.shuffled()
-    }
+    private val probabilities = listOf(
+        50f,    // Nothing: 50%
+        0.1f,   // Legendary: 0.5%
+        32f,    // Common: 30%
+        16f,    // Rare: 15%
+        1.9f    // Epic: 4.5%
+    )
 
-    companion object {
-        private const val BASE_SPIN_SPEED = 0.1f
-        private const val MAX_SPIN_SPEED = 20f
-        private const val MIN_SPIN_SPEED = 0.1f
-        private const val SPIN_SLOWDOWN_FACTOR = 0.99f
-        private const val PAUSE_DURATION = 4000L
-        const val SPIN_COST = 100
-    }
+    private var isAnimating = true
+    private var animationRunnable: Runnable? = null
 
-    private fun calculatePrize(rotationAngle: Float, pattern: List<WheelSegment>): WheelSegment {
-        val segmentAngle = 360f / pattern.size
-        val normalizedAngle = (270 - (rotationAngle % 360)) % 360
-        val segmentIndex = (normalizedAngle / segmentAngle).toInt()
-        return pattern[segmentIndex]
-    }
-
-    private fun deductSpinCost(userViewModel: UserViewModel): Boolean {
-        userViewModel.currentUser.value?.let { currentUser ->
-            val wallet = currentUser.details?.wallet ?: return false
-            if (!wallet.isSolvable(SPIN_COST, 0)) return false
-            wallet.debitCoins(SPIN_COST)
-            val updatedDetails = currentUser.details.copy(wallet = wallet)
-            val updatedUser = currentUser.copy(details = updatedDetails)
-            userViewModel.updateUser(updatedUser)
-            userViewModel.setCurrentUser(updatedUser)
-            return true
+    // Range mapping for probabilities
+    private val probabilityRanges = run {
+        var currentValue = 0
+        probabilities.map { probability ->
+            val range = (probability * 10).toInt()
+            val start = currentValue
+            currentValue += range
+            start until currentValue
         }
-        return false
     }
 
-    @Composable
-    private fun PrizeText(prize: WheelSegment) {
-        val infiniteTransition = rememberInfiniteTransition()
-        val scale by infiniteTransition.animateFloat(
-            initialValue = 1f,
-            targetValue = if (prize == WheelSegment.LEGENDARY) 1.2f else 1.1f,
-            animationSpec = infiniteRepeatable(
-                animation = tween(500),
-                repeatMode = RepeatMode.Reverse
+    private var rotation = 0f
+    private var isSpinning = false
+    private var targetRotation = 0f
+    private var currentRotation = 0f
+    private var spinStartTime = 0L
+    private var spinDuration = 0L
+    private var lastUpdateTime = System.currentTimeMillis()
+    private val idleRotationSpeed = -0.2f
+    private var pauseAfterSpin = false
+    private var pauseStartTime = 0L
+    private val pauseDuration = 2000L
+    private var expectedSegment = 0 // To track expected segment for logging
+
+    private data class Segment(val name: String, val color: Int)
+
+    internal fun stopAnimation() {
+        isAnimating = false
+        animationRunnable?.let { removeCallbacks(it) }
+    }
+
+    internal fun startAnimation() {
+        isAnimating = true
+        animationRunnable?.let { post(it) }
+    }
+
+    private fun getSegmentAtPointer(rotation: Float): Int {
+        val segmentAngle = 360f / segments.size
+        val normalizedAngle = (270 - (rotation % 360)) % 360
+        return (normalizedAngle / segmentAngle).toInt()
+    }
+
+    private fun getSegmentName(index: Int): String {
+        return segments[index].name
+    }
+
+    private fun determineTargetSegment(): Int {
+        val randomValue = (0..999).random()
+        Log.d(TAG, "Generated random value: $randomValue")
+
+        for (i in probabilityRanges.indices) {
+            if (randomValue in probabilityRanges[i]) {
+                Log.d(TAG, "Random value $randomValue falls in range for segment: ${getSegmentName(i)}")
+                return i
+            }
+        }
+        return 0
+    }
+
+    init {
+        // Log probability ranges for debugging
+        probabilityRanges.forEachIndexed { index, range ->
+            Log.d(TAG, "Probability range for ${getSegmentName(index)}: ${range.first}-${range.last}")
+        }
+
+        animationRunnable = object : Runnable {
+            override fun run() {
+                if (!isAnimating) return
+
+                val currentTime = System.currentTimeMillis()
+                val deltaTime = currentTime - lastUpdateTime
+                lastUpdateTime = currentTime
+
+                if (isSpinning) {
+                    val progress = (currentTime - spinStartTime).toFloat() / spinDuration
+                    if (progress >= 1f) {
+                        rotation = targetRotation % 360
+                        isSpinning = false
+                        pauseAfterSpin = true
+                        pauseStartTime = currentTime
+
+                        // Log final landing segment
+                        val finalSegment = getSegmentAtPointer(rotation)
+                        Log.d(TAG, "Wheel stopped on segment: ${getSegmentName(finalSegment)}")
+                        Log.d(TAG, "Expected segment was: ${getSegmentName(expectedSegment)}")
+                        if (finalSegment != expectedSegment) {
+                            Log.e(TAG, "ERROR: Wheel landed on wrong segment!")
+                        }
+                    } else {
+                        // Quartic ease-out for more gradual slowdown
+                        val easeOut = 1f - (1f - progress) * (1f - progress) * (1f - progress) * (1f - progress)
+                        rotation = currentRotation + (targetRotation - currentRotation) * easeOut
+                    }
+                } else if (pauseAfterSpin) {
+                    if (currentTime - pauseStartTime >= pauseDuration) {
+                        pauseAfterSpin = false
+                    }
+                } else {
+                    rotation += idleRotationSpeed * (deltaTime / 16.67f)
+                    if (rotation <= -360f) rotation += 360f
+                }
+
+                invalidate()
+                if (isAnimating) {
+                    postDelayed(this, 16)
+                }
+            }
+        }
+
+        post(animationRunnable!!)
+    }
+
+    override fun onDraw(canvas: Canvas) {
+        super.onDraw(canvas)
+
+        val centerX = width / 2f
+        val centerY = height / 2f
+        val radius = minOf(width, height) / 2f * 0.9f
+
+        canvas.save()
+        canvas.rotate(rotation, centerX, centerY)
+
+        val segmentAngle = 360f / segments.size
+        segments.forEachIndexed { index, segment ->
+            paint.color = segment.color
+
+            val startAngle = index * segmentAngle
+            canvas.save()
+            paint.style = Paint.Style.FILL
+            canvas.drawArc(
+                centerX - radius,
+                centerY - radius,
+                centerX + radius,
+                centerY + radius,
+                startAngle,
+                segmentAngle,
+                true,
+                paint
             )
+            canvas.restore()
+
+            // Draw segment text
+            paint.color = Color.WHITE
+            paint.textSize = radius * 0.15f
+            paint.textAlign = Paint.Align.CENTER
+
+            val middleAngleRadians = Math.toRadians(startAngle + segmentAngle / 2.0)
+
+            val textRadius = radius * 0.6f
+            val textX = centerX + cos(middleAngleRadians).toFloat() * textRadius
+            val textY = centerY + sin(middleAngleRadians).toFloat() * textRadius
+
+            canvas.save()
+            val textAngle = startAngle + segmentAngle / 2
+            canvas.rotate(textAngle + 90, textX, textY)
+
+            val bounds = Rect()
+            paint.getTextBounds(segment.name, 0, segment.name.length, bounds)
+            val textHeight = bounds.height()
+
+            canvas.drawText(
+                segment.name,
+                textX,
+                textY + textHeight / 2,
+                paint
+            )
+            canvas.restore()
+        }
+
+        canvas.restore()
+
+        // Draw the pointer
+        paint.color = Color.RED
+        paint.style = Paint.Style.FILL
+
+        val pointerPath = Path()
+        val pointerWidth = radius * 0.1f
+        val pointerHeight = radius * 0.1f
+
+        pointerPath.moveTo(centerX, centerY - radius + pointerHeight*1.5f)
+        pointerPath.lineTo(centerX - pointerWidth/2, centerY - radius - pointerHeight/2)
+        pointerPath.lineTo(centerX + pointerWidth/2, centerY - radius - pointerHeight/2)
+        pointerPath.close()
+
+        canvas.drawPath(pointerPath, paint)
+    }
+
+    fun spin() {
+        if (!isSpinning && !pauseAfterSpin) {
+            val currentSegment = getSegmentAtPointer(rotation)
+            Log.d(TAG, "Starting spin from segment: ${getSegmentName(currentSegment)}")
+
+            val targetSegment = determineTargetSegment()
+            expectedSegment = targetSegment
+            Log.d(TAG, "Target segment is: ${getSegmentName(targetSegment)}")
+
+            // Calculate required rotation to reach target segment center
+            val segmentAngle = 360f / segments.size
+            var requiredRotation = ((targetSegment - currentSegment) * segmentAngle)
+
+            // Calculate where we'll land within the target segment
+            val landingAngle = (270 - ((rotation - requiredRotation) % 360)) % 360
+            val angleWithinSegment = landingAngle % segmentAngle
+
+            val shouldTriggerNearMiss = Math.random() < 0.7f  // 70% chance for near-miss
+            var randomOffset = 0f
+
+            when (targetSegment) {
+                0 -> { // Nothing segment
+                    if (shouldTriggerNearMiss) {
+                        // For Nothing, we want to get close to the right edge (near Legendary)
+                        val maxOffsetRight = segmentAngle - angleWithinSegment
+                        // Use 90-95% of the distance to get very close to Legendary
+                        randomOffset = maxOffsetRight * (0.90f + (Math.random() * 0.05f).toFloat())
+                        Log.d(TAG, "Triggering near-miss on Nothing segment (close to Legendary)")
+                    } else {
+                        // Normal random offset within safe bounds
+                        val maxOffsetLeft = angleWithinSegment
+                        val maxOffsetRight = segmentAngle - angleWithinSegment
+                        val safeOffset = minOf(maxOffsetLeft, maxOffsetRight) * 0.9f
+                        randomOffset = (Math.random() * 2 * safeOffset - safeOffset).toFloat()
+                    }
+                }
+                2 -> { // Common segment
+                    if (shouldTriggerNearMiss) {
+                        // For Common, we want to get close to the left edge (near Legendary)
+                        val maxOffsetLeft = angleWithinSegment
+                        // Use only 5-10% of the distance to get very close to Legendary
+                        randomOffset = -maxOffsetLeft * (0.90f + (Math.random() * 0.05f).toFloat())
+                        Log.d(TAG, "Triggering near-miss on Common segment (close to Legendary)")
+                    } else {
+                        // Normal random offset within safe bounds
+                        val maxOffsetLeft = angleWithinSegment
+                        val maxOffsetRight = segmentAngle - angleWithinSegment
+                        val safeOffset = minOf(maxOffsetLeft, maxOffsetRight) * 0.9f
+                        randomOffset = (Math.random() * 2 * safeOffset - safeOffset).toFloat()
+                    }
+                }
+                else -> {
+                    // For other segments, use normal random offset within safe bounds
+                    val maxOffsetLeft = angleWithinSegment
+                    val maxOffsetRight = segmentAngle - angleWithinSegment
+                    val safeOffset = minOf(maxOffsetLeft, maxOffsetRight) * 0.9f
+                    randomOffset = (Math.random() * 2 * safeOffset - safeOffset).toFloat()
+                }
+            }
+
+            requiredRotation += randomOffset
+
+            // Ensure we're always rotating clockwise
+            if (requiredRotation <= 0) requiredRotation += 360
+
+            // Add 6-7 full rotations for effect
+            val fullRotations = (6..7).random() * 360
+            targetRotation = rotation - (requiredRotation + fullRotations)
+
+            Log.d(TAG, "Landing angle within segment: $angleWithinSegment")
+            Log.d(TAG, "Applied offset: $randomOffset degrees")
+            Log.d(TAG, "Required rotation: $requiredRotation degrees")
+            Log.d(TAG, "Total rotation with full spins: ${requiredRotation + fullRotations} degrees")
+            Log.d(TAG, "Near-miss triggered: $shouldTriggerNearMiss")
+
+            currentRotation = rotation
+            spinStartTime = System.currentTimeMillis()
+            spinDuration = 10000 // 4 seconds spin
+            isSpinning = true
+        }
+    }
+}
+
+@Composable
+fun GamblingScreen(
+    onWheelViewCreated: ((WheelView) -> Unit)? = null
+) {
+    var wheelView by remember { mutableStateOf<WheelView?>(null) }
+
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .testTag("gambling_screen"),
+        contentAlignment = Alignment.Center
+    ) {
+        AndroidView(
+            modifier = Modifier
+                .size(300.dp)
+                .testTag("wheel_view"),
+            factory = { context ->
+                WheelView(context).also {
+                    wheelView = it
+                    onWheelViewCreated?.invoke(it)
+                }
+            }
         )
 
-        AnimatedVisibility(
-            visible = true,
-            enter = fadeIn() + expandVertically()
+        Button(
+            onClick = {
+                wheelView?.spin()
+            },
+            modifier = Modifier
+                .size(90.dp)
+                .clip(CircleShape)
+                .testTag("spin_button"),
+            colors = ButtonDefaults.buttonColors(
+                containerColor = androidx.compose.ui.graphics.Color.Red
+            )
         ) {
             Text(
-                text = prize.displayName,
-                modifier = Modifier
-                    .scale(scale)
-                    .padding(top = 32.dp),
-                fontSize = when(prize) {
-                    WheelSegment.LEGENDARY -> 40.sp
-                    WheelSegment.EPIC -> 33.sp
-                    else -> 22.sp
-                },
+                "SPIN",
+                fontSize = 16.sp,
+                fontWeight = FontWeight.Bold,
                 textAlign = TextAlign.Center,
-                color = when(prize) {
-                    WheelSegment.NOTHING -> Color.Gray
-                    WheelSegment.COMMON -> Color.Green
-                    WheelSegment.RARE -> Color(0xFF0066FF)
-                    WheelSegment.EPIC -> Color(0xFF800080)
-                    WheelSegment.LEGENDARY -> Color(0xFFFFD700)
-                }
-            )
-        }
-    }
-
-    @Composable
-    fun GamblingScreenContent(
-        userViewModel: UserViewModel = viewModel(factory = UserViewModel.Factory)
-    ) {
-        var rotationAngle by remember { mutableStateOf(0f) }
-        var isSpinning by remember { mutableStateOf(false) }
-        var isPaused by remember { mutableStateOf(false) }
-        var spinSpeed by remember { mutableStateOf(BASE_SPIN_SPEED) }
-        var currentPrize by remember { mutableStateOf<WheelSegment?>(null) }
-        var currentPattern by remember { mutableStateOf(generateNewPattern()) }
-        var showInsufficientFunds by remember { mutableStateOf(false) }
-
-        val currentUser by userViewModel.currentUser.collectAsState()
-        val currentBalance = currentUser?.details?.wallet?.getCoins() ?: 0
-
-        LaunchedEffect(Unit) {
-            while(true) {
-                delay(16)
-                when {
-                    isSpinning -> {
-                        rotationAngle = (rotationAngle + spinSpeed) % 360f
-                        spinSpeed *= SPIN_SLOWDOWN_FACTOR
-
-                        if(spinSpeed <= MIN_SPIN_SPEED) {
-                            isSpinning = false
-                            isPaused = true
-                            spinSpeed = 0f
-                            currentPrize = calculatePrize(rotationAngle, currentPattern)
-                            delay(PAUSE_DURATION)
-                            isPaused = false
-                            spinSpeed = BASE_SPIN_SPEED
-                            currentPattern = generateNewPattern()
-                        }
-                    }
-                    isPaused -> {}
-                    else -> {
-                        currentPrize = null
-                        rotationAngle = (rotationAngle + BASE_SPIN_SPEED) % 360f
-                    }
-                }
-            }
-        }
-
-        Scaffold { paddingValues ->
-            Box(
-                modifier = Modifier.fillMaxSize(),
-                contentAlignment = Alignment.Center
-            ) {
-                Text(
-                    text = "Balance: $currentBalance coins",
-                    modifier = Modifier
-                        .align(Alignment.TopCenter)
-                        .padding(top = 16.dp),
-                    color = Color.White
-                )
-
-                PrizeWheel(
-                    rotationAngle = rotationAngle,
-                    segmentPattern = currentPattern
-                )
-
-                currentPrize?.let { prize ->
-                    Box(
-                        modifier = Modifier
-                            .align(Alignment.TopCenter)
-                            .padding(top = 64.dp)
-                            .fillMaxWidth(),
-                        contentAlignment = Alignment.Center
-                    ) {
-                        PrizeText(prize)
-                    }
-                }
-
-                if (showInsufficientFunds) {
-                    Text(
-                        text = "Insufficient funds!",
-                        color = Color.Red,
-                        modifier = Modifier
-                            .align(Alignment.Center)
-                            .padding(bottom = 100.dp)
-                    )
-                }
-
-                Button(
-                    modifier = Modifier
-                        .align(Alignment.BottomCenter)
-                        .padding(bottom = 32.dp),
-                    enabled = !isSpinning && !isPaused,
-                    onClick = {
-                        if (deductSpinCost(userViewModel)) {
-                            isSpinning = true
-                            spinSpeed = MAX_SPIN_SPEED
-                            currentPrize = null
-                            showInsufficientFunds = false
-                        } else {
-                            showInsufficientFunds = true
-                        }
-                    }
-                ) {
-                    Text(
-                        when {
-                            isSpinning -> "Spinning..."
-                            isPaused -> currentPrize?.displayName ?: "Getting Result..."
-                            else -> "SPIN! ($SPIN_COST coins)"
-                        }
-                    )
-                }
-            }
-        }
-    }
-
-    @Composable
-    private fun PrizeWheel(
-        modifier: Modifier = Modifier,
-        rotationAngle: Float,
-        segmentPattern: List<WheelSegment>
-    ) {
-        Canvas(
-            modifier = Modifier.size(300.dp)
-        ) {
-            val center = size.width / 2
-            val radius = size.width / 2
-            val segmentAngle = 2f * PI.toFloat() / segmentPattern.size
-
-            rotate(degrees = rotationAngle) {
-                segmentPattern.forEachIndexed { index, segment ->
-                    val startAngle = index * segmentAngle
-                    val path = Path().apply {
-                        moveTo(center, center)
-                        val rect = Rect(
-                            left = 0f,
-                            top = 0f,
-                            right = size.width,
-                            bottom = size.height
-                        )
-                        arcTo(
-                            rect = rect,
-                            startAngleDegrees = Math.toDegrees(startAngle.toDouble()).toFloat(),
-                            sweepAngleDegrees = Math.toDegrees(segmentAngle.toDouble()).toFloat(),
-                            forceMoveTo = false
-                        )
-                        lineTo(center, center)
-                        close()
-                    }
-
-                    val color = when(segment) {
-                        WheelSegment.NOTHING -> WheelColors.Nothing
-                        WheelSegment.COMMON -> WheelColors.Common
-                        WheelSegment.RARE -> WheelColors.Rare
-                        WheelSegment.EPIC -> WheelColors.Epic
-                        WheelSegment.LEGENDARY -> WheelColors.Legendary
-                    }
-
-                    drawPath(path = path, color = color)
-                    drawPath(path = path, color = Color.White, style = Stroke(width = 2f))
-                }
-
-                drawCircle(color = Color.White, radius = radius, style = Stroke(width = 4f))
-                drawCircle(color = Color.White, radius = radius * 0.1f)
-            }
-
-            drawLine(
-                color = Color.Black,
-                start = Offset(center, center),
-                end = Offset(center, center - radius + 10f),
-                strokeWidth = 6f,
-                cap = StrokeCap.Round
-            )
-
-            drawCircle(
-                color = Color.Black,
-                radius = 8f,
-                center = Offset(center, center)
+                modifier = Modifier.wrapContentSize()
             )
         }
     }
